Rename and re-stucture services
