+++ balancer-js/src/sdk.module.ts
@@ -1,21 +1,17 @@
-import {
-    BalancerNetworkConfig,
-    BalancerSdkConfig,
-    BalancerSdkSorConfig,
-} from '../types';
-import { SwapsService } from '../swapsService';
-import { RelayerService } from '../relayerService';
+import { BalancerNetworkConfig, BalancerSdkConfig } from './types';
+import { Swaps } from './modules/swaps/swaps.module';
+import { Relayer } from './modules/relayer/relayer.module';
 import { SOR } from '@balancer-labs/sor';
-import { SorFactory } from '../sor/sorFactory';
-import { BALANCER_NETWORK_CONFIG } from '../constants/contracts';
+import { SorFactory } from './sor/sorFactory';
+import { BALANCER_NETWORK_CONFIG } from './constants/contracts';
 import { JsonRpcProvider, Provider } from '@ethersproject/providers';
-import { createSubgraphClient, SubgraphClient } from '../subgraph/subgraph';
+import { createSubgraphClient, SubgraphClient } from './subgraph/subgraph';
 
 export class BalancerSDK {
     public readonly network: BalancerNetworkConfig;
     public readonly rpcUrl: string;
-    public readonly swaps: SwapsService;
-    public readonly relayer: RelayerService;
+    public readonly swaps: Swaps;
+    public readonly relayer: Relayer;
     public readonly sor: SOR;
     public readonly provider: Provider;
     public readonly subgraphClient: SubgraphClient;
@@ -33,8 +29,8 @@ export class BalancerSDK {
             this.subgraphClient
         );
 
-        this.swaps = new SwapsService(this.network, this.sor, this.provider);
-        this.relayer = new RelayerService(this.swaps, this.rpcUrl);
+        this.swaps = new Swaps(this.network, this.sor, this.provider);
+        this.relayer = new Relayer(this.swaps);
     }
 
     private getNetworkConfig(config: BalancerSdkConfig): BalancerNetworkConfig {
